+++ src/networks/types/IOLITE.js
@@ -1,5 +1,5 @@
-import tokens from '@/tokens/tokens-iolite.json';
-import contracts from '@/contracts/contract-abi-iolite.json';
+import tokens from '@/_generated/tokens/tokens-iolite.json';
+import contracts from '@/_generated/contracts/contract-abi-iolite.json';
 import iolite from '@/assets/images/networks/iolite.svg';
 
 export default {