+++ src/wallets/bip44/index.js
@@ -5,7 +5,8 @@ import {
   BITBOX02,
   SECALOT,
   KEEPKEY,
-  MNEMONIC
+  MNEMONIC,
+  COOLWALLET
 } from './walletTypes';
 import ledgerPaths from './ledgerPaths';
 import trezorPaths from './trezorPaths';
@@ -14,6 +15,7 @@ import bitbox02Paths from './bitbox02Paths';
 import secalotPaths from './secalotPaths';
 import keepkeyPaths from './keepkeyPaths';
 import mnemonicPaths from './mnemonicPaths';
+import coolWalletPaths from './coolWalletPaths';
 
 export default {
   [LEDGER]: ledgerPaths,
@@ -22,5 +24,6 @@ export default {
   [BITBOX02]: bitbox02Paths,
   [SECALOT]: secalotPaths,
   [KEEPKEY]: keepkeyPaths,
-  [MNEMONIC]: mnemonicPaths
+  [MNEMONIC]: mnemonicPaths,
+  [COOLWALLET]: coolWalletPaths
 };