+++ src/wallets/hardware/index.js
@@ -4,11 +4,16 @@ import LedgerWallet from './ledger';
 import TrezorWallet from './trezor';
 import SecalotWallet from './secalot';
 import KeepkeyWallet from './keepkey';
+import BCVaultWallet from './bcvault';
+import CoolWallet from './coolwallet';
+
 export {
   BitBoxWallet,
   BitBox02Wallet,
   LedgerWallet,
   TrezorWallet,
   SecalotWallet,
-  KeepkeyWallet
+  KeepkeyWallet,
+  BCVaultWallet,
+  CoolWallet
 };