+++ src/wallets/hardware/ledger/index.js
@@ -18,6 +18,7 @@ import {
 } from '../../utils';
 import { Misc } from '@/helpers';
 import errorHandler from './errorHandler';
+import Vue from 'vue';
 
 const NEED_PASSWORD = false;
 const OPEN_TIMEOUT = 10000;
@@ -89,10 +90,10 @@ class ledgerWallet {
       const signedChainId = calculateChainIdFromV(tx.v);
       if (signedChainId !== networkId)
         throw new Error(
-          'Invalid networkId signature returned. Expected: ' +
-            networkId +
-            ', Got: ' +
-            signedChainId,
+          Vue.$i18n.t('errorsGlobal.invalid-network-id-sig', {
+            got: signedChainId,
+            expected: networkId
+          }),
           'InvalidNetworkId'
         );
       return getSignTransactionObject(tx);