+++ tests/unit/specs/src/layouts/GettingStarted/components/WhatIfILoseMyKeysOrPassword/WhatIfILoseMyKeysOrPassword.spec.js
@@ -23,6 +23,11 @@ describe('WhatIfILoseMyKeysOrPassword.vue', () => {
     });
   });
 
+  afterEach(() => {
+    wrapper.destroy();
+    wrapper = null;
+  });
+
   it('should render correct contents', () => {
     expect(
       wrapper.vm.$el