+++ vueConfig/connections.js
@@ -57,5 +57,7 @@ module.exports = [
   'https://protocol-api.aave.com:443',
   'wss://bridge.walletconnect.org:443',
   'http://localhost:8178',
-  'ws://127.0.0.1:8178'
+  'ws://127.0.0.1:8178',
+  'https://localhost:1991', // server for bcvault
+  'wss://bridge.walletconnect.org:443'
 ];